+++ rigidBody.cpp
@@ -1,27 +1,27 @@
 #include "rigidBody.hpp"
 #include <iostream>
 
-float springConstant = 100.0f; // N/m - initialized with default value
-float e = 1; //coefficent of restitution should be 1 for perfectly elastic collision
+double springConstant = 100.0; // N/m - initialized with default value
+double e = 1; //coefficent of restitution should be 1 for perfectly elastic collision
 
 // RigidBody constructor
-RigidBody::RigidBody(float radius, float mass, 
-                     std::array<float, 2> position,
-                     std::array<float, 2> velocity)
+RigidBody::RigidBody(double radius, double mass, 
+                     std::array<double, 2> position,
+                     std::array<double, 2> velocity)
     : radius_(radius), position_(position), velocity_(velocity)
 {
     // Initialize mass and calculate inverse mass
     mass_ = mass;
-    invMass_ = (mass > 0.0f) ? (1.0f / mass) : 0.0f;
+    invMass_ = (mass > 0.0) ? (1.0 / mass) : 0.0;
     
     // Initialize other members
-    acceleration_ = {0.0f, 0.0f};
-    force_ = {0.0f, 0.0f};
+    acceleration_ = {0.0, 0.0};
+    force_ = {0.0, 0.0};
 }
 
 // RigidBody method implementation
-void RigidBody::numericalIntegration(float dt) {
-    if (mass_ == 0.0f) return;
+void RigidBody::numericalIntegration(double dt) {
+    if (mass_ == 0.0) return;
 
     // Numerical Integration. 
     // We are using Semi-Implicit Euler Method for integration. So, we need to update the velocity first and then the position.
@@ -41,8 +41,8 @@ void RigidBody::numericalIntegration(float dt) {
     position_[1] += velocity_[1] * dt;
 
     // Reset force
-    force_[0] = 0.0f;
-    force_[1] = 0.0f;
+    force_[0] = 0.0;
+    force_[1] = 0.0;
 }
 
 // TimeManager method implementations
@@ -71,12 +71,12 @@ bool TimeManager::physicsTime()
 
 // Free function implementations
 bool areColliding(const RigidBody& body1, const RigidBody& body2) {
-    std::array<float, 2> pos1 = body1.getPosition();
-    std::array<float, 2> pos2 = body2.getPosition();
-    float dx = pos1[0] - pos2[0];
-    float dy = pos1[1] - pos2[1];
-    float distSq = dx * dx + dy * dy;
-    float rSum = body1.getRadius() + body2.getRadius();
+    std::array<double, 2> pos1 = body1.getPosition();
+    std::array<double, 2> pos2 = body2.getPosition();
+    double dx = pos1[0] - pos2[0];
+    double dy = pos1[1] - pos2[1];
+    double distSq = dx * dx + dy * dy;
+    double rSum = body1.getRadius() + body2.getRadius();
     return distSq < rSum * rSum;
 }
 
@@ -84,8 +84,8 @@ void calculateForce(RigidBody& body1, RigidBody& body2) {
     // Hooke's Law for Elastic Collision
     // F = -k * (x - x0)
     // Where k is the spring constant and x0 is the equilibrium position.
-    std::array<float, 2> pos1 = body1.getPosition();
-    std::array<float, 2> pos2 = body2.getPosition();
+    std::array<double, 2> pos1 = body1.getPosition();
+    std::array<double, 2> pos2 = body2.getPosition();
     
     body1.setForceX(springConstant * (pos1[0] - pos2[0]));
     body1.setForceY(springConstant * (pos1[1] - pos2[1]));
@@ -102,26 +102,26 @@ void calculateImpulse(RigidBody&body1, RigidBody& body2)
     auto vel2 = body2.getVelocity();
 
     //find the collision vector between the 2
-    float dx = pos2[0] - pos1[0];
-    float dy = pos2[1] - pos1[1];
+    double dx = pos2[0] - pos1[0];
+    double dy = pos2[1] - pos1[1];
 
     //normalize it
-    float distance = sqrt(dx*dx + dy*dy); // find distance between the 2 so we can nromalize
+    double distance = sqrt(dx*dx + dy*dy); // find distance between the 2 so we can nromalize
     
     if (distance == 0)
     {
         return;
     }
 
-    float nx = dx/distance;
-    float ny = dy/distance;
+    double nx = dx/distance;
+    double ny = dy/distance;
 
     //now find velocity relative to each other
     auto v_rel_x = vel2[0] - vel1[0];
     auto v_rel_y = vel2[1] - vel1[1];
 
     //find the nomral by doing a dot product
-    float v_rel_normal = (v_rel_x * nx) + (v_rel_y * ny);
+    double v_rel_normal = (v_rel_x * nx) + (v_rel_y * ny);
 
     //if they are already seperating or sliding no need to apply an oppsite impulse over and over.
     if (v_rel_normal >= 0)
@@ -134,8 +134,8 @@ void calculateImpulse(RigidBody&body1, RigidBody& body2)
     //j = --------------------------------
     //      (1/m1 + 1/m2)
 
-    float numerator = -1 * (1 + e) * v_rel_normal;
-    float invMassSum = body1.getInvMass() + body2.getInvMass();
+    double numerator = -1.0 * (1.0 + e) * v_rel_normal;
+    double invMassSum = body1.getInvMass() + body2.getInvMass();
 
     //if both inv masses are 0 its imovable vs imovable. this means sim broke, 
     //no need to break it more with calculations
@@ -144,7 +144,7 @@ void calculateImpulse(RigidBody&body1, RigidBody& body2)
         return;
     }
 
-    float j = numerator/invMassSum; //impulse
+    double j = numerator/invMassSum; //impulse
 
     //now calculate new velcoties and set them
     body1.setVelocityX(vel1[0] - ((j * body1.getInvMass()) * nx));
@@ -160,24 +160,24 @@ void resolveCollision( RigidBody& b1, RigidBody& b2)
     //get vector between the 2 centers.
     //NOTE this is an arrow pointing from b1 to b2
     //so in the final calc we subtract from b1 and add to b2 along the normal.
-    std::array<float, 2> pos1 = b1.getPosition();
-    std::array<float, 2> pos2 = b2.getPosition();
-    float dx = pos2[0] - pos1[0];
-    float dy = pos2[1] - pos1[1];
+    std::array<double, 2> pos1 = b1.getPosition();
+    std::array<double, 2> pos2 = b2.getPosition();
+    double dx = pos2[0] - pos1[0];
+    double dy = pos2[1] - pos1[1];
 
     //now get distance that currently is AND  min Distance that is legal
-    float distance = sqrt(dx*dx + dy*dy);
-    float minDistance = b1.getRadius() + b2.getRadius();
+    double distance = sqrt(dx*dx + dy*dy);
+    double minDistance = b1.getRadius() + b2.getRadius();
 
     //if phasing
     if (distance < minDistance) {
-        float nx, ny;
+        double nx, ny;
         // calculate the Normal (Unit Vector)
         // use distance to make sure the vector length is 1.0
         if (distance == 0)
         {
             nx = 0.001;
-            ny = 0;
+            ny = 0.0;
         }
         else
         {
@@ -186,10 +186,10 @@ void resolveCollision( RigidBody& b1, RigidBody& b2)
         }
        
         // calculate the overlap
-        float overlap = minDistance - distance;
+        double overlap = minDistance - distance;
 
         //find total inv mass sum
-        float invMassSum = b1.getInvMass() + b2.getInvMass();
+        double invMassSum = b1.getInvMass() + b2.getInvMass();
 
         //we cooked if this happens
         if(invMassSum == 0)
@@ -199,15 +199,15 @@ void resolveCollision( RigidBody& b1, RigidBody& b2)
 
         //now move them along the normal based on their respective masses.
         //find out how much they move by mulaiplying the overlap to the individual InvMass/sum Invmass
-        float move1 = overlap*(b1.getInvMass()/invMassSum);
-        float move2 = overlap*(b2.getInvMass()/invMassSum);
+        double move1 = overlap*(b1.getInvMass()/invMassSum);
+        double move2 = overlap*(b2.getInvMass()/invMassSum);
 
         //then find components along x and y by multiplying this to the normal
-        float dx1 = nx*move1;
-        float dx2 = nx*move2;
+        double dx1 = nx*move1;
+        double dx2 = nx*move2;
 
-        float dy1 = ny*move1;
-        float dy2 = ny*move2;
+        double dy1 = ny*move1;
+        double dy2 = ny*move2;
 
         b1.setPositionX(pos1[0]-dx1);
         b1.setPositionY(pos1[1]-dy1);
@@ -218,21 +218,21 @@ void resolveCollision( RigidBody& b1, RigidBody& b2)
     }
 }
 
-void borderCheck(RigidBody& body1, std::array<float, 2> border) {
-    float radius = body1.getRadius();
-    std::array<float, 2> pos = body1.getPosition();
-    std::array<float, 2> vel = body1.getVelocity();
+void borderCheck(RigidBody& body1, std::array<double, 2> border) {
+    double radius = body1.getRadius();
+    std::array<double, 2> pos = body1.getPosition();
+    std::array<double, 2> vel = body1.getVelocity();
 
     for (int i = 0; i < 2; ++i) {
         // Lower bound (0 + radius)
         if (pos[i] < radius) {
             pos[i] = radius; 
-            vel[i] *= -1;
+            vel[i] *= -1.0;
         }
         // Upper bound (border - radius)
         else if (pos[i] > border[i] - radius) {
             pos[i] = border[i] - radius;
-            vel[i] *= -1;
+            vel[i] *= -1.0;
         }
     }
     
@@ -244,7 +244,7 @@ void runPhysics(RigidBody& body1, RigidBody& body2, const TimeManager& TIME)
 {
     // Checking if colliding
     // If colliding, calculate the force between the two bodies.
-    std::array<float, 2> border = {800, 600};
+    std::array<double, 2> border = {800.0, 600.0};
     borderCheck(body1,border);
     borderCheck(body2,border);
     if (areColliding(body1, body2)) {
@@ -252,10 +252,10 @@ void runPhysics(RigidBody& body1, RigidBody& body2, const TimeManager& TIME)
         // calculateForce(body1, body2);
         calculateImpulse(body1,body2);
 
-        // std::array<float, 2> pos1 = body1.getPosition();
-        // std::array<float, 2> vel1 = body1.getVelocity();
-        // std::array<float, 2> pos2 = body2.getPosition();
-        // std::array<float, 2> vel2 = body2.getVelocity();
+        // std::array<double, 2> pos1 = body1.getPosition();
+        // std::array<double, 2> vel1 = body1.getVelocity();
+        // std::array<double, 2> pos2 = body2.getPosition();
+        // std::array<double, 2> vel2 = body2.getVelocity();
         // std::cout<<"Body 1 Position: "<<pos1[0]<<", "<<pos1[1]<<std::endl;
         // std::cout<<"Body 1 Velocity: "<<vel1[0]<<", "<<vel1[1]<<std::endl;
         // std::cout<<"Body 2 Position: "<<pos2[0]<<", "<<pos2[1]<<std::endl;
